Fix review comment: duplicated struct __swap_ranges_fn
